220223 Refactoring - create Take & SaveScreenshot.js
